refactor: move the list Markdown codec into its own module
list.ts had grown past a thousand lines covering the schema, plugins,
config, commands, queries and the Markdown codec. The package already
splits table/ and link/ this way.

- list-markdown.ts takes parse and serialize, ~240 lines.
- list-attrs.ts takes the vocabulary both halves need, so neither imports
  the other and no cycle appears.
- markdown() only ever read listNodeName, so it now takes that rather than
  the whole config.

Pure move: list.ts loses 267 lines and gains the imports plus one call.
